Consolidate simple CLI commands
